chore(client/): Extract code in utils/collection-form to separate files
